Switch git clone branch to develop
As the documentation is currently presented, it recommends using master. Doing so leads to some files missing in later steps.
Changing this to cloning the develop branch instead of the master branch, allows the installation process to go further (though it still currently runs into trouble, see for reference https://pastebin.com/ggegKjeG )
I am not sure if the desired solution here is to change the instruction to use develop, or to fix the master branch so it includes the correct files, but here is a pull request in case the first option is the correct one.